another try

+++ lib/tasks/clear_data.rake
@@ -1,12 +1,11 @@
 namespace :clear_data do
-  CCPHA_DATABASE_INSTANCE_NAME = "Heal Cities - CCPHA"
 
   desc "clear out the user data"
   task clear_ccpha_contacts: :environment do
 
     dbi_ccpha = Heal::DatabaseInstance.find_by(instance_name: CCPHA_DATABASE_INSTANCE_NAME)
 
-    dbi_ccpha.contacts.destroy_all
+    dbi_ccpha.contacts.destroy_all if dbi_ccpha.present?
 
   end
